sdk/python/storage/keeper_secrets_manager_storage/storage_gcp_kms/requirements_gcp.txt
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-kms==3.3.0
google-crc32c==1.6.0
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
importlib_metadata==8.6.1
keeper-secrets-manager-core==16.6.6
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
requests==2.32.3
rsa==4.9
urllib3==2.3.0
zipp==3.21.0
